Allow data scripts to run on the foreground as well.

+++ ds/scripts/data-get.sh
@@ -1,7 +1,11 @@
 #!/bin/bash -x
 
-nohup_out=/host/logs/nohup-get.out
-rm -f $nohup_out
-nohup nice /var/www/data/get/all.sh > $nohup_out &
-sleep 1
-tail -f $nohup_out
+if [[ $1 == 'nobg' ]]; then
+    nice /var/www/data/get/all.sh
+else
+    nohup_out=/host/logs/nohup-get.out
+    rm -f $nohup_out
+    nohup nice /var/www/data/get/all.sh > $nohup_out &
+    sleep 1
+    tail -f $nohup_out
+fi
